Split API into env and util files
