--- .env
@@ -1,2 +1,0 @@
-PORT=3000
-API_PORT=3001
