+++ .env.defaults
@@ -1,0 +1,5 @@
+NODE_ENV=production
+APP_TITLE=React Calendar API
+PORT=3000
+API_PORT=3001
+REL_SNAP_DIR=../snapshots
