+++ .gitignore
@@ -11,6 +11,7 @@
 
 # production
 /build
+.env
 
 # misc
 .DS_Store
