+++ api/api.js
@@ -6,63 +6,23 @@ const du = require('du')
 const fs = require('fs')
 const url = require('url')
 const path = require('path')
-const winston = require('winston')
 const pidusage = require('pidusage')
 const puppeteer = require('puppeteer')
 const handlebars = require('handlebars')
-const expressWinston = require('express-winston')
 const expAutoSan = require('express-autosanitizer')
 
-// Globals
-const APP_TITLE = 'React Calendar API'
-const REL_SNAP_DIR = '../snapshots'
-const ABS_SNAP_DIR = path.join(__dirname, REL_SNAP_DIR)
+// Locals
+const config = require('./env').config()
+const util = require('./util')
+const logger = util.createLogger()
+const expressLogger = util.createExpressLogger()
 
-const isDebugging = () => {
-  const debugging_mode = {
-    headless: false,
-    slowMo: 250,
-    devtools: true,
-  }
-  return process.env.NODE_ENV === 'debug' ? debugging_mode : {}
-}
-
-const logger = winston.createLogger({
-  level: 'info',
-  format: winston.format.combine(
-    winston.format.colorize(),
-    winston.format.json()
-  ),
-  defaultMeta: { service: 'api' },
-});
-
-
-// If we're not in production then log to the `console` with the format:
-// `${info.level}: ${info.message} JSON.stringify({ ...rest }) `
-if (process.env.NODE_ENV !== 'production') {
-  logger.add(new winston.transports.Console({
-    format: winston.format.simple()
-  }));
-} else {
-    logger.add(new winston.transports.File({ filename: 'combined.log' }));
-    logger.add(new winston.transports.File({ filename: 'error.log', level: 'error' }));
-}
-
+// Instantiation
 let browser, page;
 const app = express();
-
 app.use(cors());
 app.use(expAutoSan.all)
-
-app.use(expressWinston.logger({
-  transports: [new winston.transports.Console()],
-  format: winston.format.json(),
-  meta: true, // optional: control whether you want to log the meta data about the request (default to true)
-  msg: "HTTP {{req.method}} {{req.url}} {{req.headers}} {{req.connection}}", // optional: customize the default logging message. E.g. "{{res.statusCode}} {{req.method}} {{res.responseTime}}ms {{req.url}}"
-  expressFormat: true, // Use the default Express/morgan request formatting. Enabling this will override any msg if true. Will only output colors with colorize set to true
-  colorize: false, // Color the text and status code, using the Express/morgan color palette (text: gray, status: default green, 3XX cyan, 4XX yellow, 5XX red).
-  ignoreRoute: function (req, res) { return false; } // optional: allows to skip some log messages based on request and/or response
-}));
+app.use(expressLogger);
 
 
 /**
@@ -77,10 +37,10 @@ app.get('/stats', async (req, res) => {
   try{
     const proc = browser.process()
     puppeteerStats = await pidusage(proc.pid)
-    snapDirSize = await du(ABS_SNAP_DIR)
+    snapDirSize = await du(config.ABS_SNAP_DIR)
 
     // Todo: also list file names? Maybe at a different endpoint...
-    const snapFiles = await fs.readdirSync(ABS_SNAP_DIR)
+    const snapFiles = await fs.readdirSync(config.ABS_SNAP_DIR)
     numSnapFiles = snapFiles.length
   } catch(err){
     logger.error('stats :: ', err)
@@ -98,32 +58,12 @@ app.get('/stats', async (req, res) => {
 
 
 /**
- * Builds a filename given a url.parse() object.
- * Translates query params into a suitable file name.
- * Might be better to use md5() to generate names in the future, but that
- * would obfuscate filenames and make development a bit harder.
- **/
-function buildFilename(parsedUrl){
-  const parsedPath = path.parse(parsedUrl.pathname)
-  const parsedQuery = parsedUrl.query.replace(/=/gi, '_')
-  const fileName = `${parsedPath.name}-${parsedQuery}${parsedPath.ext}`
-  return fileName
-}
-
-function buildFilepath(parsedUrl){
-  const fileName = buildFilename(parsedUrl)
-  const filePath = path.join(__dirname, REL_SNAP_DIR, `calendar${fileName}.png`)
-  return filePath
-}
-
-
-/**
  * Raw image URL
  * Serves an image after translating GET params into a filename.
  **/
 app.get('/', async (req, res) => {
   const hrstart = process.hrtime()
-  const filePath = buildFilepath(req._parsedUrl)
+  const filePath = util.buildFilepath(req._parsedUrl)
 
   if (!fs.existsSync(filePath)) {
     try {
@@ -151,36 +91,22 @@ app.get('/', async (req, res) => {
 
 
 /**
- * System startup tests
- * Ensure the environment is suitable for operating within.
- **/
-function testSystem(){
-  let success = true;
-
-  try {
-    fs.accessSync(ABS_SNAP_DIR, fs.constants.R_OK | fs.constants.W_OK)
-  } catch(err){
-    logger.error(`testSystem read/write error: ensure directory exists: ${ABS_SNAP_DIR}`)
-    success = false
-  }
-
-  return success
-}
-
-/**
  * Server instantiation
  * Opens an empty Puppeteer instance so that incoming requests
  * don't have to start so cold.
  **/
-const server = app.listen(process.env.API_PORT, async () => {
-  if(!testSystem()){
-    logger.error('Startup failure')
+const server = app.listen(config.API_PORT, async () => {
+  try {
+    await util.testSystem()
+  } catch(err) {
+    logger.error('Startup failure :: ', err)
     process.exit()
   }
 
-  browser = await puppeteer.launch(isDebugging())
+  browser = await puppeteer.launch(util.isDebugging())
   page = await browser.newPage();
-  logger.info(`API listening on http://localhost:${process.env.API_PORT}`)
+  logger.info(`NODE_ENV = ${config.NODE_ENV}`)
+  logger.info(`API listening on http://localhost:${config.API_PORT}`)
 });
 
 
@@ -188,10 +114,5 @@ const server = app.listen(process.env.API_PORT, async () => {
  * Ensure Chromium is closed and the process restarts cleanly
  * for both user and nodemon exit signals
  **/
-async function cleanup() {
-  console.log('cleanup...')
-  !!browser && await browser.close()
-  server.close(() => process.kill(process.pid, 'SIGUSR2'))
-}
-process.once('exit', async () => await cleanup())
-process.once('SIGUSR2', async () => await cleanup())
+process.once('exit', async () => await util.cleanup(browser, server))
+process.once('SIGUSR2', async () => await util.cleanup(browser, server))
