+++ package.json
@@ -9,6 +9,7 @@
     "cors": "^2.8.5",
     "date-fns": "^2.11.1",
     "dotenv": "^8.2.0",
+    "dotenv-defaults": "^1.1.1",
     "du": "^1.0.0",
     "express": "^4.17.1",
     "express-autosanitizer": "^1.0.1",
